requirements.txt
streamlit
langchain_community
langchain
torch
pymupdf
python-dotenv
faiss-cpu
fastapi
sentence-transformers